Add console controller and fix services error
