+++ src/middleware/errorHandler.js
@@ -1,5 +1,6 @@
 const errorHandler = (err, req, res, next) => {
   console.error(err.stack)
+  console.log('TEST')
   res.status(500).send(err.message)
 }
