+++ src/services/brands.js
@@ -14,7 +14,7 @@ class BrandsService {
       await dbConnection.query(insertQuery);
       return "Added succesfully!";
     } catch(err) {
-      throw new Error(err.messsage);
+      throw new Error(err);
     } 
     
   }
@@ -34,7 +34,7 @@ class BrandsService {
       await dbConnection.query(updateQuery);
       return "Updated succesfully!";
     } catch(err) {
-      throw new Error(err.messsage);
+      throw new Error(err);
     } 
   }
 
@@ -50,7 +50,7 @@ class BrandsService {
       await dbConnection.query(removeQuery);
       return "Removed succesfully!";
     } catch(err) {
-      throw new Error(err.messsage);
+      throw new Error(err);
     } 
   }
   
@@ -64,7 +64,7 @@ class BrandsService {
       const rows = await dbConnection.query(selectQuery);
       return rows;
     } catch(err) {
-      throw new Error(err.messsage);
+      throw new Error(err);
     } 
     
   }
@@ -80,7 +80,7 @@ class BrandsService {
       const brand = result[0];
       return brand || {};
     } catch(err) {
-      throw new Error(err.messsage);
+      throw new Error(err);
     } 
     
   }
